+++ lib/default.nix
@@ -1,6 +1,6 @@
-{ args, ... }:
+{ systems, version, lib, util, inputs, channels, path, files }:
 rec
 {
-  build = import ./build.nix args;
-  map = import ./map.nix args;
+  build = import ./build.nix { inherit systems version lib util inputs channels path files; };
+  map = import ./map.nix { inherit systems lib inputs files; };
 }