+++ modules/nix/default.nix
@@ -1,6 +1,7 @@
-{ util, pkgs, ... }:
+{ config, lib, util, pkgs, ... }:
 let
   inherit (util) map;
+  hardware = config.hardware;
 in rec
 {
   imports = [ ./cachix.nix ];
@@ -31,6 +32,7 @@ in rec
       trustedUsers = [ "root" "@wheel" ];
 
       # Additional Features
+      maxJobs = if (builtins.hasAttr "cores" hardware) then hardware.cores else 4;
       extraOptions = "experimental-features = nix-command flakes recursive-nix";
       systemFeatures = [ "kvm" "recursive-nix" ];
     };